moved amazon and ebay api logic to separate modules
